gh-653: (refactor) rename pkg get_meta -> getmeta
The underscore has been annoying the linter (and me) for quite a while.
