+++ graphqlapi/local/resolver.go
@@ -14,11 +14,11 @@ package local
 
 import (
 	local_get "github.com/creativesoftwarefdn/weaviate/graphqlapi/local/get"
-	local_get_meta "github.com/creativesoftwarefdn/weaviate/graphqlapi/local/get_meta"
+	local_getmeta "github.com/creativesoftwarefdn/weaviate/graphqlapi/local/getmeta"
 )
 
 // Resolving local GraphQL queries
 type Resolver interface {
 	local_get.Resolver
-	local_get_meta.Resolver
+	local_getmeta.Resolver
 }
